% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacent_municipalities.R
\docType{data}
\name{adjacent_municipalities}
\alias{adjacent_municipalities}
\title{List of adjacent municipalities}
\format{
A list with 1117 elements
}
\source{
Fill in later!
}
\usage{
adjacent_municipalities
}
\description{
A convenient representation of the adjacency matrix of all the municipalities
in Colombia.Each item is a vector containing the codes of all municipalities that are
adjacent to the municipality whose code is the name of that list item.
}
\examples{
adjacent_municipalities[1]
names(adjacent_municipalities[1])

}
\keyword{datasets}
