% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{bootstrap_F1}
\alias{bootstrap_F1}
\title{Bootstrap test for threshold effect following Hansen (1999, JOE)}
\usage{
bootstrap_F1(X, Y, n_bootstrap = 500, q_lower = 0.05, q_upper = 0.95,
  n_grid = 180)
}
\arguments{
\item{X}{Panel data for x. Assumes that each row is an individual and each
column is a time period.}

\item{Y}{Panel data for y. Assumes that each row is an individual and each
column is a time period.}

\item{n_bootstrap}{Number of bootstrap draws}

\item{q_lower}{Quantile of x above which we will search for the threshold}

\item{q_upper}{Quantile of of x below which we will search for the threshold}

\item{n_grid}{Number of grid points between \code{q_lower} and \code{q_lower}
over which we will search for the the threshold}
}
\value{
List with bootstrap results:
\itemize{
   \item \code{gamma_hat} Estimated threshold location
   \item \code{gamma_seq} Grid of values considered for threshold location
   \item \code{S0} Sum of squared residuals for restricted model with no threshold
   \item \code{coef0} Estimated regression coefficients for restricted model with no
   \item \code{e_tilde} Residuals from the restricted model with no threshold
   threshold
   \item \code{e_gamma_hat} Residuals for threshold model with gamma set equal to
   gamma_hat
   \item \code{S1_gamma_hat} Sum of squared residuals for threshold model with gamma
   set equal to gamma_hat
   \item \code{S1} Vector corresponding to gamma_seq. Each element is the sum of squared
   residuals for a thresdhold model with that particular value of gamma.
   \item \code{F1} Value of the F-test statistic for the input dataset
   \item \code{F1_boot} Vector of values for F-test statistic in each of the
   bootstrap samples
   \item \code{pvalue} Bootstrap p-value for test of the null hypothesis that there is no
   threshold, calculated from F1 and F1_boot
   \item \code{n_i} Number of individuals in the panel
   \item \code{n_T} Number of time periods in the panel
 }
}
\description{
Bootstrap test for threshold effect following Hansen (1999, JOE)
}
\examples{
test_data <- dgp_threshold()
results <- bootstrap_F1(test_data$X, test_data$Y)
}
