% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{dgp_threshold}
\alias{dgp_threshold}
\title{Generate test dataset to test for threshold effects}
\usage{
dgp_threshold(d = 1, n_i = 200, n_t = 17, gamma = 6, b1 = 1,
  b2 = -0.1)
}
\arguments{
\item{d}{Size of the threshold effect}

\item{n_i}{Number of individuals}

\item{n_t}{Number of time periods}

\item{gamma}{Location of threshold}

\item{b1}{Coefficient on x}

\item{b2}{Coefficient on x-squared}
}
\value{
List with panel data X and Y formatted appropriately for bootstrap_F1
}
\description{
Generate test dataset to test for threshold effects
}
\examples{
test_data <- dgp_threshold()
}
