% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{gamma_CI}
\alias{gamma_CI}
\title{Confidence set for estimated threshold location following Hansen (1999, JOE)}
\usage{
gamma_CI(results, alpha = 0.05)
}
\arguments{
\item{results}{List of results from bootstrap_F1}

\item{alpha}{Confidence level. Defaults to 0.05}
}
\value{
Vector of values for gamma from \code{results$gamma_seq} that do not
give a significantly different sum of squared residuals than \code{results$gamma_hat}
based on a likelihood ratio test with level alpha
}
\description{
Confidence set for estimated threshold location following Hansen (1999, JOE)
}
\examples{
test_data <- dgp_threshold()
results <- bootstrap_F1(test_data$X, test_data$Y)
gamma_CI(results)
}
