% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel.R
\docType{data}
\name{panel}
\alias{panel}
\title{Panel dataset}
\format{
A tibble with 14,127 rows and 13 variables:
\describe{
  \item{municipality}{municipality code}
  \item{year}{year}
  \item{V_cum}{count of cumulative paramilitary violence (murders) in
  a given municipality up to and including the specified year}
  \item{V_flow}{count of paramilitary violence (murders) in a given
  municipality during the specified year}
  \item{placeboV_cum}{same as V_cum but guerrilla rather than paramilitary
  violence}
  \item{placeboV_flow}{same as V_flow but guerrilla rather than paramilitary
  violence}
  \item{D_AS}{Count of people displaced by from a given municipality in a
  given year, as reported by AS}
  \item{D_CODHES}{Same as D_AS, but as reported by CODHES rather than AS}
  \item{D_RUV}{Same as D_AS, but as reported by RUV rather than AS}
  \item{D_CEDE}{Same as D_AS, but as reported by CEDE rather than AS}
  \item{D_JYP}{Same as D_AS, but as reported by JYP rather than AS}
  \item{popn1993}{Population of the municipality in 1993 from census data}
  \item{lag_V_flow}{One year lag of V_flow}
}
}
\source{
Fill in later!
}
\usage{
panel
}
\description{
Annual data for paramilitary violence along with five measures of population
displacement, as measured by multiple agencies, in each municipality of Colombia
between 1996 and 2012
}
\keyword{datasets}
