% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{plot_gamma}
\alias{plot_gamma}
\title{Plot confidence set for estimated threshold location following Hansen (1999, JOE)}
\usage{
plot_gamma(results, alpha = 0.05)
}
\arguments{
\item{results}{List of results from bootstrap_F1}

\item{alpha}{Confidence level. Defaults to 0.05}
}
\value{
Plots the results of the likelihood ratio test from \code{gamma_CI}.
Each point is a test statistic corresponding to a particular value of
\code{results$gamma_seq}. The red line gives the critical value for a particular
choice of \code{alpha}, which defaults to 0.05. Points below the red line
correspond to values of gamma that cannot be rejected.
}
\description{
Plot confidence set for estimated threshold location following Hansen (1999, JOE)
}
\examples{
test_data <- dgp_threshold()
results <- bootstrap_F1(test_data$X, test_data$Y)
plot_gamma(results)
}
